feat: Keyboard navigation state is truly global
Previously there was an entity called `KeyborgState` that was stored in
a global file scope variable called `__state`. This broke keyboard
navigation state when multiple keyborg instances were used in different
bundles.

This PR removes `KeyborgState` and uses `KeyborgCore` to hold the
keyboard navigation state. This is because there can only be one
`KeyborgCore` instance per window anyway

https://github.com/microsoft/keyborg/blob/acf66e1053f18d3863c5a2a73c4340afcfa4b78f/src/Keyborg.ts#L312-L327

The result is that the keyboard navigation state will be truly global
and depends on the one core instance that is attached to the window.

Fixes #45